rename docker subpackage
