--- python/packages/autogen-ext/src/autogen_ext/code_executor/docker_command_line_code_executor/__init__.py
@@ -1,3 +1,0 @@
-from ._impl import DockerCommandLineCodeExecutor
-
-__all__ = ["DockerCommandLineCodeExecutor"]
